FINALLY! fixed generation, works with koch too
I was applying the transformation to the translation vector before
drawing it.

Now i apply it only after, and before passing it to the next iteration,
as should be.

Now both sierpinski and koch work, but not fern. I'll have to check if
there's a problem with the fern definition or again with the algoritm.

EDIT: the fern works too, with a change in the zoom factor.